Split up ImportData module
